changed time limit to be minutes, moved turntimer to client.ui.swing.util
