+++ megamek/i18n/megamek/common/options/messages.properties
@@ -1,7 +1,7 @@
 GameOptionsInfo.group.basic.displayableName=Basic Options
 GameOptionsInfo.option.friendly_fire.displayableName=Friendly fire
 GameOptionsInfo.option.friendly_fire.description=If checked,\nthe game will allow attacks against friendly units,\ncontrary to the rules.\nUnchecked by default.
-GameOptionsInfo.option.turn_timer.displayableName=Turn timer
+GameOptionsInfo.option.turn_timer.displayableName=Turn timer in minutes
 GameOptionsInfo.option.turn_timer.description=If value is greater 0,\n the game will end a players current action turn after the time is up.\n deactivated by 0 value by default.
 GameOptionsInfo.option.skip_ineligable_movement.displayableName=Skip ineligible during movement
 GameOptionsInfo.option.skip_ineligable_movement.description=If checked,\nthe game will skip a unit during the movement phase if it is immobile or otherwise inactive.\nUnchecked by default.
