+++ megamek/i18n/megamek/common/options/messages_de.properties
@@ -1,7 +1,7 @@
 GameOptionsInfo.group.base.displayableName=Grundoptionen
 GameOptionsInfo.option.friendly_fire.displayableName=Friendly fire
 GameOptionsInfo.option.friendly_fire.description=Diese Option deaktivieren,\num bei Verwendung der Schaltfläche 'nächstes Ziel' eigene Einheiten zu überspringen.\nVoreinstellung: Deaktiviert.
-GameOptionsInfo.option.turn_timer.displayableName=Zug Zeitbegrenzung
+GameOptionsInfo.option.turn_timer.displayableName=Zug Zeitbegrenzung in Minuten
 GameOptionsInfo.option.turn_timer.description=Ist der Wert größer als 0,\n wird das Spiel die aktuelle Aktionsphase des Spielers beenden, sobald die Zeit abgelaufen ist.\n Wert 0 standardmäßig deaktiviert.
 GameOptionsInfo.option.skip_ineligable_movement.displayableName=Bewegungsunfähige Einheiten überspringen
 GameOptionsInfo.option.skip_ineligable_movement.description=Wenn aktiviert,\nso werden bewegungsunfähige Einheiten für die Bewegungsinitiative nicht berücksichtigt.\nVoreinstellung: Deaktiviert.
